moved create event and made the my event list work
There is some very poorly written code in here. For this I am sorry. It works though, so that's a plus. Changed some minor things with the user class (added an attribute that keeps track of what events the user is organizing, as well as methods to set get and add too it) added a whole bunch of random bullshit to profileFragment.java. I'm going to sleep soon. (also the my event list will not work for any events made before this is merged)